Move options_menu modules to new sub folder components/options_menu
