+++ app/frontend/components/options_menu/locale_switch.js
@@ -16,9 +16,9 @@
    along with Ezra Bible App. See the file LICENSE.
    If not, see <http://www.gnu.org/licenses/>. */
 
-const { html, waitUntilIdle } = require('../helpers/ezra_helper.js');
-const i18nController = require('../controllers/i18n_controller.js');
-const i18nHelper = require('../helpers/i18n_helper.js');
+const { html, waitUntilIdle } = require('../../helpers/ezra_helper.js');
+const i18nController = require('../../controllers/i18n_controller.js');
+const i18nHelper = require('../../helpers/i18n_helper.js');
 const locales = i18nController.getAvailableLocales();
 
 const LOCALE_SELECT_WIDTH = '170px'; // FIXME: magic number that works with jQuery UI is not reliable way for constructing stable layout
