+++ app/frontend/components/options_menu/select_option.js
@@ -16,7 +16,7 @@
    along with Ezra Bible App. See the file LICENSE.
    If not, see <http://www.gnu.org/licenses/>. */
 
-const { waitUntilIdle } = require('../helpers/ezra_helper.js');
+const { waitUntilIdle } = require('../../helpers/ezra_helper.js');
 
 class SelectOption extends HTMLSelectElement {
   constructor() {
