+++ app/frontend/controllers/app_controller.js
@@ -35,7 +35,7 @@ const TextController = require("./text_controller.js");
 const VerseContextController = require("./verse_context_controller.js");
 const TabSearch = require("../components/tab_search/tab_search.js");
 const TabController = require("./tab_controller.js");
-const OptionsMenu = require("../components/options_menu.js");
+const OptionsMenu = require("../components/options_menu/options_menu.js");
 const NavigationPane = require("../components/navigation_pane.js");
 const TranslationComparison = require("../components/translation_comparison.js");
 const BookSelectionMenu = require("../components/book_selection_menu.js");
