+++ app/frontend/startup.js
@@ -90,9 +90,9 @@ class Startup {
 
     const fs = require('fs');
 
-    require('./components/config_option.js');
-    require('./components/select_option.js');
-    require('./components/locale_switch.js');
+    require('./components/options_menu/config_option.js');
+    require('./components/options_menu/select_option.js');
+    require('./components/options_menu/locale_switch.js');
 
     var bookSelectionMenu = fs.readFileSync('html/book_selection_menu.html');
     var tagSelectionMenu = fs.readFileSync('html/tag_selection_menu.html');
